ZSH: Install ohmyzsh using .chezmoiexternal
The script simply clones the repo + does some extra (chsh, running ZSH)
with a lot of error handling and pretty messages.

Using the toml config will also help the repo to update occasinally,
and I can now move the chsh and running zsh to be in a manual script
at the end of the "once" installation.